code: tidy residual fixes (Phase 20) — loop-var, widening cast, macro parens, UVM padding
Five remaining strict-tidy findings, fixed individually:

  - bugprone-too-small-loop-variable in
    src/object/umodule_instance.c:195 — widen `uint16_t i` to `size_t`
    so the loop var matches the upper bound (`m->nested_count` is
    size_t).  Functionally a no-op (nested_count never exceeds 16-bit
    in practice) but the analyzer is right that the contract should
    match the type.

  - bugprone-misplaced-widening-cast in src/value/uvalue.c:195 —
    rewrite `(size_t)(n + 2) < cap` as `(size_t)n + 2U < cap` so the
    addition is performed in size_t (after `n` is widened) rather than
    in `int` (then widened, with possible overflow on int-max-2).

  - bugprone-macro-parentheses in src/vm/uvm.c:48 (DISPATCH macro) —
    cannot parenthesize a `goto *expr` statement; suppress per-line
    with rationale in the surrounding comment block.

  - clang-analyzer-optin.performance.Padding on struct UVM
    (src/vm/uvm.h:115) — the optimal-pack reordering would shrink the
    struct by ~40 bytes (host) but break the 6 _Static_assert layout
    pins and scatter milestone-clustered fields across the body.
    Suppress per-line with rationale in the surrounding comment block.
    Accounts for both reported padding errors (one per TU that
    includes the header).

Closes 5 strict-tidy findings — strict-tidy now reports 0.

+++ src/object/umodule_instance.c
@@ -192,7 +192,7 @@ urbi_module_instance_create(struct UVM *vm, UModule *m)
     /* entries[1..n-1]: parallel to module->nested[].  Each gets its own
      * slice of the trailing IC region; unfilled sites have topology_gen == 0
      * (the sentinel per pre-M4 topology-generation spec §3.1). */
-    for (uint16_t i = 0U; i < m->nested_count; i++) {
+    for (size_t i = 0U; i < m->nested_count; i++) {
         UProto *p = m->nested[i];
         if (p != NULL &&
             !intern_ic_names_from_strs(vm, p->ic_count, &p->ic_names,
